add variable paths/request parameters
