+++ static/index.js
@@ -6,10 +6,10 @@ if (remove) {
 
 function onremove(ev) {
     let node = ev.target;
-    let id = node.dataset.id;
+    let userId = node.dataset.id;
     let res = new XMLHttpRequest();
 
-    res.open("DELETE", "/" + id);
+    res.open("DELETE",  "/" + userId);
     res.onload = onload;
     res.send();
 
@@ -17,6 +17,6 @@ function onremove(ev) {
         if (res.status !== 200) {
             throw new Error("Kan de tag niet verwijderen!");
         }
-        window.location = "/droppedTags";
+        window.location = "/searchLocation";
     }
 }
