+++ view/add.ejs
@@ -2,14 +2,14 @@
 <title>Match Tag - Register</title>
 <main class="add">
   <h1>Drop jouw Tag</h1>
-  <form action=/droppedTags method=post enctype=multipart/form-data> 
+  <form action="/<%= collTitle %>"  method=post enctype=multipart/form-data> 
     <label>Upload profielfoto<input name=cover type=file accept=image/* value="<%= user.cover %>"></label>
       <label>Naam<input name=name value="<%= user.name %>"></label>
     <label>Leeftijd<input type="number" name=age value="<%= user.age %>"></label>
     <label>Openingszin<textarea name=pickupline rows="1" value="<%= user.pickupline %>"></textarea></label>
     <label>Schrijf iets over jezelf<textarea name=description rows="5" value="<%= user.description %>"></textarea></label>
     <section class="bottomBtns">
-    <a href="droppedTags">Annuleer</a>
+    <a href="javascript:history.back()">Annuleer</a>
     <button>Drop Tag</button>
   </section>
   </form>
