+++ view/detail.ejs
@@ -3,9 +3,9 @@
 <header>
   <nav class="navBar">
     <ul>
-      <li class=""><a href="login"></a></li>
-      <li class=""><a href="searchLocation"></a></li>
-      <li class=""><a href="personalTag"></a></li>
+      <li class=""><a href="../login"></a></li>
+      <li class=""><a href="../searchLocation"></a></li>
+      <li class=""><a href="../personalTag"></a></li>
     </ul>
   </nav>
 </header>
