+++ view/searchLocation.ejs
@@ -25,21 +25,5 @@
               <% } %>
             </ul>
           </section>
-
-
-    <!-- <section class="midSection">
-        <div class="radar-container">
-            <div class="img-container">
-                    <a href="droppedTags">
-                <img src="/static/images/games/rdr2.jpg" alt="Red Dead Redemption 2 cover image">
-                </a>
-            </div>
-            <p>0,5 km</p>
-            <p>1,0 km</p>
-            <p>1,5 km</p>
-        </div>
-    </section>
-
-            <a class="dropTagBtn" href=/add>Drop jouw tag</a>  -->
 </main>
 <% include tail.ejs %>
